packages/backend/requirements.txt
boto3>=1.28
requests>=2.31
fastapi
uvicorn[standard]
sentence-transformers>=2.5
faiss-cpu>=1.7
langchain>=0.1.0
